+++ config/application.rb
@@ -46,7 +46,7 @@ module Leihs
     config.autoload_paths << Rails.root.join('lib')
     config.autoload_paths << Rails.root.join('database/lib')
 
-    config.middleware.use "AuditedRequests"
+    config.middleware.insert_before "ActionDispatch::ShowExceptions", "Leihs::AuditedRequests"
   end
 end
